Move post materials to material module
